+++ src/edges_analysis/analysis/__init__.py
@@ -1,7 +1,6 @@
 from . import beams
 from . import coordinates
 from . import filters
-from . import io
 from . import levels
 from . import loss
 from . import plots
@@ -10,3 +9,7 @@ from . import sky_models
 from . import tools
 
 from .levels import Level1, Level2, Level4, Level3
+
+from pathlib import Path
+
+DATA = Path(__file__).absolute().parent / "data"