+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## Version dev
+
+### Added
+- Methods for filtering on Level1 objects, and calling them consistently from Level2.
+- Store which files are fully flagged in Level2.
+- Beam class for easier handling of beam information
+- IndexModel class for defining models of spectral indices on the sphere.
+- Parallel filtering at Level2
+
+### Fixed
+- Warnings in xrfi are now suppressed, and a summary is given.
+
 ## Version 0.1.0
 
 First version compatible with new `edges-collab` set of repos.